Document the rest of the code
This brings GoodJob to 89% documented -- the remaining bits are the methods on `LogSubscriber`, which didn't seem worth documenting since they merely mirror the names of notifications they respond to, so I feel like they're pretty self documenting. These can probably be handled with one of:
- `:nodoc:` comments (which unfortunately show up in the resulting docs)
- A macro that makes a pretty simplistic comment
- `@api nodoc` comments combined with `--hide-api nodoc` in the `.yardopts` file. This will hide them completely.

This also adds the `yard-activesupport-concern` gem, which is pretty helpful for documenting methods inside `included` and `class_methods` blocks, which are Rails features that are wind up hiding their contents from Yard.