Fix start-up issue with Rails 6.1 & async mode
I'm not sure this is an acceptable fix, but this would not raise `queue_parser': undefined method `first' for "*":String (NoMethodError) anymore.

I also notice the start-up time for async mode is significantly slower than other modes. (It took more than 5 minutes to start-up).

I'm not sure it is related to this fix or any other factor of rails 6.1.