update Gemfile.lock to latest dependencies
Via running `bundle update` and committing results.

Some would advise not having Gemfile.lock checked into repo for gem source in the first place. Eg https://yehudakatz.com/2010/12/16/clarifying-the-roles-of-the-gemspec-and-gemfile/  https://github.com/radar/guides/issues/14

But if you are going to have it checked in, it's probably good to periodically update it so dev and test environments are against latest dependency releases that meet requirements.