Draft logs for job
This draft allows me to have an idea of what the logging module
will look like. It consists of the execution id, the logger info
and the logger debug. Those three data, could be added into a
model and would have a belong_to relationship with the Execution
 model on one end; and on the other end, the Excution would have a
has_one relationship with this model. And then, at each execution,
a record of this model would be saved in the database.