Ignore ActiveJob::DeserializationError when discarding jobs
Previously, discarding a job would attempt to deserialize the job
arguments and crash. We load the job to attempt to run the `#instrument`
method, but unfortunately this means that the job could not be
discarded.

This commit adds a flag to `Execution#active_job`,
`ignore_deserialization_errors`, which says "I want to load this job but
I don't mind if the arguments cannot be deserialized".

There are only two places currently where this method is called, and the
other one (`retry_job`) will still throw errors when attempting to retry
a job in this state. But now, when discarding, the errors are completely
ignored and the job can be discarded.