Avoid including unnecessary pg_locks references when counting jobs
Previously, the executed query for in particular the `succeeded` scope included
a join against `pg_locks`.  This is not particularly useful, and unfortunately
triggers evaluation of that join criterion, with noticable expense as job volume
grows.

This change modifies the jobs filter logic to only include those joins in the
necessary locations (`running`) and when loading records for practical
enumeration.