Flush STDOUT buffer after writing logs to STDOUT
When running good_job CLI from the console and logging to STDOUT the
log messages will be displayed right away since ruby sets sync to true
by default, on STDOUT when run from an interactive shell.

When the good_job CLI is started by foreman or in a docker container
the shell is non interactive and therefore ruby sets sync to false by
default, causing the log messages to be buffered and not shown until
STDOUT flushes the buffer (on exit for example).

In this patch we check to see if any of the log subscribers are writing
to STDOUT and if they are and $stdout.sync == false we call flush on
$stdout to show the log message right away.