If logging to STDOUT set $stdout.sync = true
When running the CLI from non-interactive terminals such as foreman or
docker, the logs were not being displayed immediatly because ruby sets
$stdout.sync = false for these types of terminals by default. This leads
to inconsistent behaviour between interactive and non-interactive
terminals for the CLI application. By explicitly setting $stdout.sync =
true when launching the CLI application we can ensure that STDOUT
logging is consistent in both types of terminals.