Automatically send status notifications to systemd (#1029)
* First pass on systemd integration

This adds a somewhat naive `SystemdService` class that will send appropriate notifications to systemd or do nothing if it doesn't detect that the process is being run by systemd. It pretty much just notifies that the process is ready, is stopping, and actively notifies the watchdog while it's running. It doesn't reach into the actual job system to check that things are OK, or hook into ActiveSupport notifications to tell systemd about other events (like restarting/reloading).

Fixes #1027.

* Log using notifications

* Add some docs

* Handle graceful shutdown notifications better

* Update Sorbet typing information

* Systemd errors should probably always log

* Add tests

* Add example systemd configuration file

* Make Sorbet happy

Stubbed constants are very not cool as far as Sorbet is concerned, so the way I wrote these tests before broke the linter. OTOH, trying to figure out how to make this acceptable did get some slightly nicer test setup.

* Skip socket tests on JRuby

* Move vendored sd_notify into lib/

* Add note about test skipping

It's good to know why these are skipped so we can stop skipping them in the future if the JRuby issue is fixed (or someone figures out a workaround).