"Force" discard jobs that are already running/runaway to prevent retry (#1073)
* Force discard jobs

Allows a new action to a running job: force discard. This will mark the
jobs as discarded (with a force discarded error message). Note that the
job will continue to run on the thread - that part is outside the scope
of this pr - but it will not be retried if the job fails. This can be
useful to stop a runaway job from being continuously restarted/retried.

* Add tests

* Lint and wrap promises with Rails executor

---------

Co-authored-by: Ben Sheldon [he/him] <bensheldon@gmail.com>