GoodJob::Capsule#idle?
Instead of pushing the logic of idle all the way up to the CLI I thought
it would be better in the GoodJob::Capsule class.

I've created a `#idle?` method that does this check based on the last
time a scheduler in the GoodJob::Capsule has run a job.

I ran into an interesting thing that if I check the configuration object
in the initialization of the GoodJob::Capsule it causes a crash because
rails is not initialized yet.  The configuation object can't access the
rails config.

I instead changed it so that in the initialize of the GoodJob::Capsule
the `shutdown_on_idle_enabled` flag is set to false - meaning never
shutdown on idle. Then in `#start` the GoodJob::Capsule checks the
config and sets the flag so that #idle? method will work.