Over wrote something by mistake
I had a stale file open and overwrote the check of the event.  This was
after I had pushed up all the code.

This also made me realize that, now one can't set less than a 10 second
idle time out.  The CLI will always stay up for at least 10 seconds.  As
mentioned in a previous issue, this may cause a support request when
someone sets a idle time out of 1 second but the CLI stays up for 10
seconds then shuts down.

In real life I can't imagine that someone would want to have a job
server around for less than 10 seconds. Maybe this just needs to be
surfaced in the documentation for the command line option.

In any case, I've added code so that the wait will take the
shutdown_on_idle as the time out to check the event, or it will take the
default of 10 seconds.

Also in my testing I realized that the `executed_at` was being set in
every loop because of the `increment_empty_executions`, this meant that
the time out would never happen during actual execution.

I've removed the setting of the `executed_at` in this function as
technically nothing executed.