Add spec to cover PgHero basic auth
We want to be sure PgHero is behind basic auth. This spec validates that
once the correct environment variables are set, PgHero will require
basic authentication.

We have to use the rack_test driver so we can use `basic_authorize` to
set the headers, rack_test is the default in RSpec/Capybara, but we have
explicitly set the driver just in case.