Ensure before actions run on root route (#1482)
* Ensure before actions run on root route

When using the recommended method for injecting a before filter into GoodJob the initial root route ignores that filter. In cases where someone is using that before filter to apply security they may want to return a non-301 response code for users who don't pass the authentication check.

For example someone may want to obfuscate that they're using GoodJob by returning a 404 from all GoodJob routes if the user is not logged in. Before this change users would receive the redirect from the root route, then get the 404, essentially "leaking" that the site is running GoodJob.

* Add query params to root route redirect.

Live polling relies on a query string param with hits the root route. Ensure this param survives the redirect.

* Move redirect action to jobs controller

---------

Co-authored-by: Ben Sheldon [he/him] <bensheldon@gmail.com>