Manually bump tapioca / regenerate rbi
Dependabot does not do this because it resolves with Ruby 3.0 from the gemspec
and tapioca now only supports 3.1

This unblocks prism 1.0 which is required for the ruby-lsp to update to the latest version.
Currently it runs into https://github.com/Shopify/ruby-lsp/issues/2630, causing a restart loop.

To make sure this does the right thing on older rubies, I added a step to print the resolved lockfile on CI