Adjust indexes on `good_jobs.finished_at` to better support job cleanup (#1677)
* Create a partial index on `good_jobs.finished_at` where those values are NOT NULL

This will better support the implementation of `cleanup_preserved_jobs`.  The
existing partial index here also requires that `retried_good_job_id` be NULL,
but that column was effectively written out in https://github.com/bensheldon/good_job/pull/1430

* Remove the extraneous partial index on `good_jobs.finished_at`

This is functionally redundant with the partial index that is only conditional
on `finished_at`, because `retried_good_job_id` is actually ignored.  There
doesn't seem any real point in continuing to pay the insert overhead for this
index going forward.

---------

Co-authored-by: Michael Fowler <michael@delivereasy.co.nz>